third_party/blink/web_tests/platform/generic/external/wpt/css/selectors/invalidation/target-pseudo-in-has-expected.txt

This is a testharness.js-based test.
FAIL CSS Selectors Invalidation: :target pseudo-class in :has() argument assert_equals: parent1 should be green on fragment click expected "rgb(0, 128, 0)" but got "rgb(128, 128, 128)"
Harness: the test ran to completion.
